+++ go.mod
@@ -1,9 +1,8 @@
-module api_date_conversion
+module github.com/kotatabe/api_date_conversion
 
 go 1.18
 
 require (
-	github.com/go-chi/chi v1.5.4
-	github.com/najeira/jpholiday v0.0.0-20201205051200-308d6dc577e8
-	golang.org/x/xerrors v0.0.0-20220411194840-2f41105eb62f
+	github.com/go-chi/chi v1.5.4 // indirect
+	golang.org/x/xerrors v0.0.0-20220411194840-2f41105eb62f // indirect
 )
